% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plots.R
\name{beta_plot}
\alias{beta_plot}
\title{Plot a beta distribution}
\usage{
beta_plot(
  shape1,
  shape2,
  cv = 0.5,
  tail = TRUE,
  limits = c(0, 1),
  color = "steelblue"
)
}
\description{
Plot a beta distribution
}
