% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plots.R
\name{chisquare_plot}
\alias{chisquare_plot}
\title{Plot a chi-squared distribution}
\usage{
chisquare_plot(
  df = 2,
  cv = 2,
  tails = "greater",
  limits = c(0, 10),
  color = "steelblue"
)
}
\arguments{
\item{cv}{the critical value(s) to shade.}

\item{limits}{the limits of the plot.}

\item{color}{the color to fill in the area.}
}
\description{
Plot a chi-squared distribution
}
\seealso{
\code{\link[stats:Chisquare]{stats::dchisq()}}
}
