% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA606-package.R
\docType{package}
\name{DATA606-package}
\alias{DATA606-package}
\title{R Package to support DATA606 Statistics and Probabilty for Data Analytics.}
\description{
DATA606 Statistics and Probabilty for Data Analytics
}
\author{
\email{jason@bryer.org}
}
\keyword{package}
\keyword{shiny}
