% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA606-package.R
\name{demo}
\alias{demo}
\title{Demonstrates R Functionality}
\usage{
demo(topic, echo = FALSE, ask = FALSE, ...)
}
\arguments{
\item{topic}{the topic which should be demonstrated, given as a
    \link{name} or literal character string, or a character string,
    depending on whether \code{character.only} is \code{FALSE} (default)
    or \code{TRUE}.  If omitted, the list of available topics is
    displayed.}

\item{echo}{a logical.  If \code{TRUE}, show the \R input when sourcing.}

\item{ask}{a logical (or \code{"default"}) indicating if
    \code{\link{devAskNewPage}(ask = TRUE)} should be called before
    graphical output happens from the demo code.  The value
    \code{"default"} (the factory-fresh default) means to ask if
    \code{echo == TRUE} and the graphics device appears to be
    interactive.  This parameter applies both to any currently opened
    device and to any devices opened by the demo code.  If this is
    evaluated to \code{TRUE} and the session is \link{interactive}, the
    user is asked to press RETURN to start.}
}
\description{
This overwrites the \code{\link{demo}} function in the \code{utils} package
setting by setting \code{ask=FALSE} and \code{echo=FALSE}.
}
