% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plots.R
\name{distribution_plot}
\alias{distribution_plot}
\title{Plots a distribution}
\usage{
distribution_plot(
  fun,
  cv,
  tails = c("no", "two.sided", "less", "greater"),
  limits,
  color = "steelblue",
  ...
)
}
\arguments{
\item{fun}{the density function (e.g. dnorm)}

\item{cv}{the critical value(s) to shade.}

\item{limits}{the limits of the plot.}

\item{color}{the color to fill in the area.}

\item{...}{other parameters passed to \code{fun}.}
}
\description{
Plots a distribution
}
