% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plots.R
\name{F_plot}
\alias{F_plot}
\title{Plot a F distribution.}
\usage{
F_plot(
  df1 = 1,
  df2 = 5,
  cv = 2,
  tails = "greater",
  limits = c(0, 10),
  color = "steelblue"
)
}
\arguments{
\item{cv}{the critical value(s) to shade.}

\item{limits}{the limits of the plot.}

\item{color}{the color to fill in the area.}

\item{cf}{critical value (or F-statistic).}
}
\description{
Plot a F distribution.
}
\seealso{
\code{\link[stats:Fdist]{stats::df()}}
}
