% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IS606-package.R
\docType{package}
\name{IS606-package}
\alias{IS606-package}
\title{R Package to support IS606 Statistics and Probabilty for Data Analytics.}
\description{
IS606 Statistics and Probabilty for Data Analytics
}
\author{
\email{jason@bryer.org}
}
\keyword{package}
\keyword{shiny}

