% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lab.R
\name{lab}
\alias{lab}
\title{Opens a lab.}
\usage{
lab(l, new_file = paste0(getwd(), "/", Sys.info()["user"], "-Lab", lab,
  ".Rmd"))
}
\arguments{
\item{l}{the lab name (see \{code{\link{getLabs}} for available labs).}

\item{new_file}{where the lab template will be copied to.}
}
\description{
Opens a lab.
}

