% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plots.R
\name{normal_plot}
\alias{normal_plot}
\title{Plot a normal distribution.}
\usage{
normal_plot(
  mean = 0,
  sd = 1,
  cv = c(-1, 1),
  tails = "no",
  limits = c(mean - 3 * sd, mean + 3 * sd),
  color = "steelblue"
)
}
\arguments{
\item{mean}{mean of the distribution.}

\item{sd}{standard deviation of the distribution.}

\item{cv}{the critical value(s) to shade.}

\item{limits}{the limits of the plot.}

\item{color}{the color to fill in the area.}
}
\description{
Plots a normal distribution with the area within the given cv shaded.
The probability (i.e. area) under the normal curve for the given cv
is also given.
}
\examples{
	normal_plot()
	normal_plot(cv=c(-2,2))
	normal_plot(cv=c(-3,3))
	normal_plot(tails=TRUE)
}
