% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalPlot.R
\name{normalPlot}
\alias{normalPlot}
\title{Plot a normal distribution.}
\usage{
normalPlot(mean = 0, sd = 1, bounds = c(-1, 1), tails = FALSE)
}
\description{
Plots a nomral distribution with the area within the given bounds shaded.
The probability (i.e. area) under the normal curve for the given bounds
is also given.
}
\seealso{
http://www.statmethods.net/advgraphs/probability.html
}

