% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plots.R
\name{t_plot}
\alias{t_plot}
\title{Plot a student t distribution.}
\usage{
t_plot(
  df = 10,
  cv = c(-1, 1),
  tails = "no",
  limits = c(mean - 3 * sd, mean + 3 * sd),
  color = "steelblue"
)
}
\arguments{
\item{cv}{the critical value(s) to shade.}

\item{limits}{the limits of the plot.}

\item{color}{the color to fill in the area.}
}
\description{
Plot a student t distribution.
}
