% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab.R
\name{viewLab}
\alias{viewLab}
\title{Opens the HTML page for the lab in the default browser.}
\usage{
viewLab(l)
}
\arguments{
\item{l}{the lab to open.}
}
\description{
Opens the HTML page for the lab in the default browser.
}
