% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA606-package.R
\name{onAttach}
\alias{onAttach}
\alias{.onAttach}
\title{Function executed when the package is loaded.}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{the library name.}

\item{pkgname}{the package name (hopefully 'DATA606').}
}
\description{
Function executed when the package is loaded.
}
